+++ components/CollectionCard.tsx
@@ -13,7 +13,7 @@ import { cn } from "@/lib/utils";
 import type { z } from "zod";
 import type { collectionsPageScheme } from "@/lib/collection";
 import { useAtomValue } from "jotai/react";
-import { personalViewModeAtom } from "@/components/personalViewSwitch";
+import { personalViewModeAtom } from "@/components/collection/personalViewSwitch";
 import { useRef, useState } from "react";
 import { useDialog } from "@/hooks/use-dialog";
