+++ components/Panel/EPList/EPListDynamic.tsx
@@ -7,7 +7,7 @@ import { useReducerAtom } from "jotai/react/utils";
 import { panelHistoryAtom, panelReducer } from "@/lib/panel";
 import { EPListItemList } from "@/components/Panel/EPList/EPListItemList";
 import { Suspense } from "react";
-import { EPItemSkeleton } from "@/components/EPItemSkeleton";
+import { EPItemSkeleton } from "@/components/Panel/EPList/EPItemSkeleton";
 import { useSubjectData } from "@/hooks/use-subject";
 
 export function EPListDynamic({
